Move some routes to routes.ts
